+++ packages/kit/src/runtime/server/utils.js
@@ -2,7 +2,6 @@ import * as devalue from 'devalue';
 import { json, text } from '../../exports/index.js';
 import { coalesce_to_error } from '../../utils/error.js';
 import { negotiate } from '../../utils/http.js';
-import { has_data_suffix } from '../../utils/url.js';
 import { HttpError } from '../control.js';
 import { fix_stack_trace } from '../shared.js';
 
@@ -52,23 +51,6 @@ export function allowed_methods(mod) {
 }
 
 /**
- * @template {'prerender' | 'ssr' | 'csr' | 'trailingSlash'} Option
- * @template {Option extends 'prerender' ? import('types').PrerenderOption : Option extends 'trailingSlash' ? import('types').TrailingSlash : boolean} Value
- *
- * @param {Array<import('types').SSRNode | undefined>} nodes
- * @param {Option} option
- *
- * @returns {Value | undefined}
- */
-export function get_option(nodes, option) {
-	return nodes.reduce((value, node) => {
-		return /** @type {any} TypeScript's too dumb to understand this */ (
-			node?.universal?.[option] ?? node?.server?.[option] ?? value
-		);
-	}, /** @type {Value | undefined} */ (undefined));
-}
-
-/**
  * Return as a response that renders the error.html
  *
  * @param {import('types').SSROptions} options
@@ -98,7 +80,7 @@ export async function handle_fatal_error(event, options, error) {
 		'text/html'
 	]);
 
-	if (has_data_suffix(new URL(event.request.url).pathname) || type === 'application/json') {
+	if (event.isDataRequest || type === 'application/json') {
 		return json(body, {
 			status
 		});