+++ packages/kit/src/runtime/shared.js
@@ -1,5 +1,5 @@
-export let assets = '';
-export let base = '';
+export { set_assets } from '$internal/paths';
+
 export let building = false;
 export let version = '';
 
@@ -11,12 +11,6 @@ export let public_env = {};
 
 /** @param {string} stack */
 export let fix_stack_trace = (stack) => stack;
-
-/** @param {{ base: string, assets: string }} paths */
-export function set_paths(paths) {
-	base = paths.base;
-	assets = paths.assets || base;
-}
 
 /** @param {boolean} value */
 export function set_building(value) {