+++ packages/kit/test/apps/basics/src/routes/snapshot/+layout.svelte
@@ -1,0 +1,5 @@
+<a href="/snapshot/a">a</a>
+<a href="/snapshot/b">b</a>
+<a href="/snapshot/c" data-sveltekit-reload>c</a>
+
+<slot />